recovery: added option table to control menu options
option table includes all options at menu. it returns the active and status state of each option.
by this way the recovery can control an option for example to disable it due to a certain case.

scheme -> 'option': {'enabled': True (active) or False (disabled), 'status': 0=ok or 1=warn or 2=error}
status 0: option has no issues and should work as expected
status 1: gives a warning to the user that this option might not work as expected
status 2: option will not work (or it makes no sense to keep it active) and is disabled

this feature is (or may be) future relevant
print table function looks kinda... well I will rework it soon