recovery: detect local git repository
it may be possible that HyperUBot got git cloned or git initialized. Using the recovery to apply updates, restore or reinstall the bot may corrupt the git history or (in worst case) break the .git completely. So warn the user with a warning line 'Directory is a local git repository' to be aware of that risk