userbot: added Properties
Properties offer a special feature to store configs, properties or similar data variables permanently to survive reboots. The main feature of this is to notify, detect, recognize changes etc. as there are cases that may require HyperUBot to stop the runtime for a certain time. So the bot can check for props and continue the work where it stopped before.