_package_manager: removed live updating
Earlier, the package manager was updating the user's message immediately if one or more modules were un/installed or if repos were added/removed to let the user knowing live what it currently does while the process is running. To to so, the message was updated using edit message each time if a subprocess finished. This however may cause a FloodWaitError to the user if too many modules are being un/installed. Due to this the message will be updated only after the whole process has finished (basic static way).